Fix: Controller 병합
