+++ src/main/java/jwp/controller/LogoutController.java
@@ -1,21 +1,21 @@
-package jwp.controller;
-
-import jwp.support.session.UserSessionUtils;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import javax.servlet.http.HttpSession;
-
-@Controller
-@RequestMapping("/user")
-public class LogoutController {
-
-    @GetMapping("/logout")
-    public String logout(HttpSession session) {
-        if (session != null) {
-            session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
-        }
-        return "redirect:/";
-    }
-}
+//package jwp.controller;
+//
+//import jwp.support.session.UserSessionUtils;
+//import org.springframework.stereotype.Controller;
+//import org.springframework.web.bind.annotation.GetMapping;
+//import org.springframework.web.bind.annotation.RequestMapping;
+//
+//import javax.servlet.http.HttpSession;
+//
+//@Controller
+//@RequestMapping("/user")
+//public class LogoutController {
+//
+//    @GetMapping("/logout")
+//    public String logout(HttpSession session) {
+//        if (session != null) {
+//            session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
+//        }
+//        return "redirect:/";
+//    }
+//}
